requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
requests
cloudscraper
python-dotenv
plotly
